Added 3 new collections & improved index
Added new collections as per request: meshing, AI, and multiphysics.
Products in these categories are subject to change and may be developed
further.

Also added an index to the index of all the products as a big list in
alphabetical order which will make searching for what you're interested
in significantly easier.